App.tsxで使っていたhookをカスタムフックに切り出した
