+++ src/App.tsx
@@ -1,20 +1,11 @@
-import React, { useState } from 'react'
+import React from 'react'
 import { menuData } from './lib/mocks/menu-data'
-import { Menu } from './lib/types/menu'
 import style from './App.module.scss'
 import TrackCafeLayout from './components/TrackCafeLayout'
+import { useMenu } from './hooks/useMenu'
 
 const App = () => {
-  const [menus, setMenus] = useState(menuData)
-  const updateMenus = (targetMenu: Menu) => {
-    const updatedMenus = menus.map((menu) => {
-      if (menu.id === targetMenu.id) {
-        return { ...menu, numberOfOrders: menu.numberOfOrders + 1 }
-      }
-      return menu
-    })
-    setMenus(updatedMenus)
-  }
+  const { menus, updateMenus } = useMenu(menuData)
   return (
     <div className={style.container}>
       <TrackCafeLayout menus={menus} updateMenus={updateMenus} />
